New language: asm

+++ python/3/main.py
@@ -8,7 +8,7 @@ def isPrime(x):
     return True
 
 BIG_NUMBER = 600851475143
-for i in range(0, math.ceil(600851475143 ** 0.5))[::-1]:
+for i in range(0, int(math.ceil(600851475143 ** 0.5)))[::-1]:
     if BIG_NUMBER % i == 0 and isPrime(i):
         print(i)
         break
